% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_feeds}
\alias{get_feeds}
\title{Get custom feeds list for an account}
\usage{
get_feeds(account_id = NULL)
}
\arguments{
\item{account_id}{your account id}
}
\value{
data frame of custom feeds
}
\description{
Get custom feeds list for an account
}
\examples{
\dontrun{
get_feeds()
}
}
