% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads.R
\name{get_leads}
\alias{get_leads}
\title{Get leads}
\usage{
get_leads(account_id = NULL, start_date = NULL, end_date = NULL)
}
\arguments{
\item{account_id}{your account id}

\item{start_date}{start date (yyyy-mm-dd)}

\item{end_date}{end date (yyyy-mm-dd)}
}
\value{
data frame of all leads
}
\description{
Get leads
}
\examples{
\dontrun{
get_leads(
account_id = "12345", start_date = "2020-11-01", end_date = "2020-11-30")
}
}
