% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_token}
\alias{set_token}
\title{Set your Leadfeeder API token}
\usage{
set_token(token = NULL)
}
\arguments{
\item{token}{API token, generated on Leadfeeder dashboard}
}
\description{
Set your Leadfeeder API token
}
\examples{
\dontrun{
set_token(token = "yourtoken")
}
}
