% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads.R
\name{get_leads}
\alias{get_leads}
\title{Get leads}
\usage{
get_leads(
  account_id = NULL,
  start_date = NULL,
  end_date = NULL,
  feed_id = NULL
)
}
\arguments{
\item{account_id}{your account id}

\item{start_date}{start date (yyyy-mm-dd)}

\item{end_date}{end date (yyyy-mm-dd)}

\item{feed_id}{custom feed id. NULL by default.}
}
\value{
data frame of all leads for all the feeds (if feed_id is NULL) or for one specific feed (if feed_id is provided)
}
\description{
Get leads
}
\examples{
\dontrun{

get_leads(
   account_id = "12345",
   start_date = "2020-11-01",
   end_date = "2020-11-30"
)
}
}
