% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits.R
\name{get_visits}
\alias{get_visits}
\title{Get all visits}
\usage{
get_visits(account_id = NULL, start_date = NULL, end_date = NULL)
}
\arguments{
\item{account_id}{your account id}

\item{start_date}{start date (yyyy-mm-dd)}

\item{end_date}{end date (yyyy-mm-dd)}
}
\value{
data frame of all visits
}
\description{
Get all visits
}
\examples{
\dontrun{

get_visits(
   account_id = "12345",
   start_date = "2020-11-01",
   end_date = "2020-11-30"
)
}
}
